Clase 2021-11-13 - Polimorfismo - Clase Figura. Clase Pila.

+++ Figuras/Cuadrado.cpp
@@ -20,5 +20,5 @@ double Cuadrado::perimetro() const
 
 double Cuadrado::apotema() const
 {
-    return lado / 2;
+    return (double)lado / 2;
 }
